added scraping module, error when fetch response != 200
